% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markPH.aipw.R
\name{markPH.aipw}
\alias{markPH.aipw}
\title{AIPW Estimation and Hypothesis Testing of Strain-Specific
Vaccine Efficacy Adjusted for Covariate Effects with Missing Causes}
\usage{
markPH.aipw(
  cmprskPHformula,
  trtpos = 1,
  strata,
  missformula,
  markformula,
  data = parent.frame(),
  VEnull = 0.3,
  maxit = 15
)
}
\arguments{
\item{cmprskPHformula}{a formula object with the response on the left of a '~'
operator, and the independent terms on the right as covariates in the
cause-specific proportional hazards regression model. The response
must be in the format of cbind(time, delta, cause), where time is the minimum
of the failure time and the censoring time, delta specifies the censoring status
(1: failure is observed; 0: right-censored), cause is the cause of failure.
If cause is missing, cause=NA.}

\item{trtpos}{the position of the treatment group indicator in the RHS of
the formula object cmprskPHformula. The default value is 1.}

\item{strata}{the column name of the strata variable in the data.}

\item{missformula}{a formula object started with a '~' operator, and the independent
terms on the right of '~' are variables used for predicting the probability
of being not missing under a logistic regression model.}

\item{markformula}{a formula object started with a '~' operator, and the independent
terms on the right of '~' are variables used for predicting the probability
of being cause j under a multinomial log-linear regression model (function
multinom() in nnet package).}

\item{data}{a data.frame with the variables.}

\item{VEnull}{the assumed VE value in the null hypothesis. The default value
is 0.3.}

\item{maxit}{Maximum number of iterations to attempt for convergence. The
default is 15.}
}
\value{
returns an object of type 'markPH.aipw'. With the following arguments:
\item{causes}{the types of causes of failure}
\item{coef.cc}{estimates of covariate coefficients using a complete-case analysis}
\item{se.cc}{estimates of standard errors of estimators for covariate coefficients
using a complete-case analysis}
\item{coef.ipw}{estimates of covariate coefficients using IPW method}
\item{se.ipw}{estimates of standard errors of estimators for covariate coefficients
using IPW method}
\item{coef}{estimates of covariate coefficients using AIPW method}
\item{se}{estimates of standard errors of estimators for covariate coefficients
using AIPW method}
\item{coef.VE}{estimates of the strain-specific vaccine efficacy to reduce
susceptibility to strain j using AIPW method, j=1,...,J}
\item{se.VE}{estimates of standard errors of estimators for the strain-specific
vaccine efficacy to reduce susceptibility to strain j using AIPW method, j=1,...,J}
\item{coef.VD}{estimates of VD using AIPW method}
\item{se.VD}{estimates of standard errors of estimators for VD using AIPW method}
\item{U1}{test statistic for testing against the alternative hypothesis
HA1: VE(j)>=VEnull with strict inequality for some 1<=j<=J}
\item{U2}{test statistic for testing against the alternative hypothesis
HA2: VE(j) is not equal to VEnull for some 1<=j<=J}
\item{U1j}{test statistic for testing against the alternative hypothesis
HAj1: VE(j)>VEnull}
\item{U2j}{test statistic for testing against the alternative hypothesis
HAj2: VE(j) is not equal to VEnull}
\item{T1}{test statistic for testing against the alternative hypothesis
HB1: VE(1)>=...>=VE(J) with strict inequality for some 1<=j<=J}
\item{T2}{test statistic for testing against the alternative hypothesis
HB2: VE(i) is not equal to VE(j) for at least one pair of i and j, 1<=i<j<=J}
\item{pval.A1}{p value for testing against the alternative hypothesis
HA1: VE(j)>=VEnull with strict inequality for some 1<=j<=J}
\item{pval.A2}{p value for testing against the alternative hypothesis
HA2: VE(j) is not equal to VEnull for some 1<=j<=J}
\item{pval.A1j}{p value for testing against the alternative hypothesis
HAj1: VE(j)>VEnull}
\item{pval.A2j}{p value for testing against the alternative hypothesis
HAj2: VE(j) is not equal to VEnull}
\item{pval.B1}{p value for testing against the alternative hypothesis
HB1: VE(1)>=...>=VE(J) with strict inequality for some 1<=j<=J}
\item{pval.B2}{p value for testing against the alternative hypothesis
HB2: VE(i) is not equal to VE(j) for at least one pair of i and j, 1<=i<j<=J}
\item{tgrid}{specifies the time grids for estimating \eqn{\lambda_{0kj}(\cdot)}
, the unspecified cause-specific baseline hazard function
for cause j and the k-th stratum.}
\item{lambda0}{estimates of \eqn{\lambda_{0kj}(\cdot)} using AIPW method}
\item{covariates}{covariates in the cause-specific proportional hazards model.}
\item{trtpos}{the position of the treatment group indicator in the RHS of
the formula object cmprskPHformula.}
\item{VEnull}{the assumed VE value in the null hypothesis.}
\item{diffsigma}{estimates of standard deviation of \eqn{\hat\alpha_j-\hat\alpha_{j-1}}, j=2,...,J.}
\item{cov.alpha}{estimated covariance matrix of \eqn{(\hat\alpha_1,\dots,\hat\alpha_J)}.}
}
\description{
Fit a stratified cause-specific proportional hazards regression model
postulates that the conditional cause-specific hazard function
for cause j for an individual in the k-th stratum with the covariate value
\eqn{z} equals
\deqn{ \lambda_{kj}(t|z)=\lambda_{0kj}(t)\exp(\beta_j^{T}z) }
for j=1,...,J and k=1, ..., K, where \eqn{\lambda_{0kj}(\cdot)}
is an unspecified cause-specific baseline hazard function
for the k-th stratum and K is the number of strata.
}
\examples{

## Example 1: Simulated competing risks data with 2 causes subject to missingness
data(sim2cs)

res.aipw <- markPH.aipw(cmprskPHformula=cbind(time,delta,cause)~z1+z2,
                        trtpos=1,
                        strata="strata",
                        missformula=~z1+A,
                        markformula=~time+z1+A,
                        data=sim2cs,
                        VEnull=0.3,
                        maxit=15)
res.aipw # print the result

## Example 2: Simulated competing risks data with 3 causes subject to missingness
data(sim3cs)

res.aipw <- markPH.aipw(cmprskPHformula=cbind(time,delta,cause)~z1+z2,
                        trtpos=1,
                        strata="strata",
                        missformula=~z1+A,
                        markformula=~time+z1+A,
                        data=sim3cs,
                        VEnull=0.3,
                        maxit=15)
res.aipw # print the result

}
\references{
(2021+)
}
\author{
Fei Heng
}
