% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markPH.aipw2.R
\name{markPH.aipw2}
\alias{markPH.aipw2}
\title{Approach 2:
AIPW Estimation and Hypothesis Testing of Strain-Specific
Vaccine Efficacy Adjusted for Covariate Effects with Missing Causes}
\usage{
markPH.aipw2(
  cmprskPHformula,
  trtpos = 1,
  strata,
  causelevels = NULL,
  missmodel = T,
  missformula,
  markformula,
  data = parent.frame(),
  VEnull = 0.3,
  maxit = 15
)
}
\arguments{
\item{cmprskPHformula}{a formula object with the response on the left of a '~'
operator, and the independent terms on the right as covariates in the
cause-specific proportional hazards regression model. The response
must be in the format of cbind(time, delta, cause), where time is the minimum
of the failure time and the censoring time, delta specifies the censoring status
(1: failure is observed; 0: right-censored), cause is the cause of failure.
If cause is missing, cause=NA.}

\item{trtpos}{the position of the treatment group indicator in the RHS of
the formula object cmprskPHformula. The default value is 1.}

\item{strata}{the column name of the strata variable in the data.}

\item{causelevels}{types of causes.}

\item{missmodel}{indicates whether a logistic regression model is applied to
predict the probability of being not missing for each type of causes under the
missing at random (MAR) assumption. If not, a logical vector is needed. TRUE:
fit a logistic regression model; FALSE: that cause is not missing, i.e. the
probability of being not missing for that cause is 1.
For example, we have three types of causes: c("cause A", "cause B", "cause C").
missmodel=c(T,T,F) means that the probability of being not missing for cause C is 1,
and the logistic regression model is fitted for cause A and cause B.
Furthermore, because cause C is not missing, we do not have to predict the probability
of being cause C while using the AIPW method.
The default value is T.}

\item{missformula}{a formula object started with a '~' operator, and the independent
terms on the right of '~' are variables used for predicting the probability
of being not missing under a logistic regression model.}

\item{markformula}{a formula object started with a '~' operator, and the independent
terms on the right of '~' are variables used for predicting the probability
of being cause j under a multinomial log-linear regression model (function
multinom() in nnet package).}

\item{data}{a data.frame with the variables.}

\item{VEnull}{the assumed VE value in the null hypothesis. The default value
is 0.3.}

\item{maxit}{Maximum number of iterations to attempt for convergence. The
default is 15.}
}
\value{
returns an object of type 'markPH.aipw2'. With the following arguments:
\item{causes}{types of causes}
\item{misscauses}{types of causes which are subject to missingness}
\item{coef.cc}{estimates of covariate coefficients using a complete-case analysis}
\item{se.cc}{estimates of standard errors of estimators for covariate coefficients
using a complete-case analysis}
\item{coef.ipw}{estimates of covariate coefficients using IPW method}
\item{se.ipw}{estimates of standard errors of estimators for covariate coefficients
using IPW method}
\item{coef.aipw}{estimates of covariate coefficients using AIPW method}
\item{se.aipw}{estimates of standard errors of estimators for covariate coefficients
using AIPW method}
\item{coef.VE}{estimates of the strain-specific vaccine efficacy to reduce
susceptibility to strain j using AIPW method, j=1,...,J}
\item{se.VE}{estimates of standard errors of estimators for the strain-specific
vaccine efficacy to reduce susceptibility to strain j using AIPW method, j=1,...,J}
\item{coef.VD}{estimates of VD using AIPW method}
\item{se.VD}{estimates of standard errors of estimators for VD using AIPW method}
\item{U1}{test statistic for testing against the alternative hypothesis
HA1: VE(j)>=VEnull with strict inequality for some j in \code{misscauses}}
\item{U2}{test statistic for testing against the alternative hypothesis
HA2: VE(j) is not equal to VEnull for some j in \code{misscauses}}
\item{U1j}{test statistic for testing against the alternative hypothesis
HAj1: VE(j)>VEnull}
\item{U2j}{test statistic for testing against the alternative hypothesis
HAj2: VE(j) is not equal to VEnull}
\item{T1}{test statistic for testing against the alternative hypothesis
HB1: VE(\code{misscauses}[1]) \eqn{\ge \dots \ge} VE(\code{misscauses}[I]) with strict inequality for some cause in \code{misscauses}}
\item{T2}{test statistic for testing against the alternative hypothesis
HB2: VE(i) is not equal to VE(j) for at least one pair of i and j in \code{misscauses}}
\item{pval.A1}{p value for testing against the alternative hypothesis
HA1: VE(j)>=VEnull with strict inequality for some j in \code{misscauses}}
\item{pval.A2}{p value for testing against the alternative hypothesis
HA2: VE(j) is not equal to VEnull for some j in \code{misscauses}}
\item{pval.A1j}{p value for testing against the alternative hypothesis
HAj1: VE(j)>VEnull}
\item{pval.A2j}{p value for testing against the alternative hypothesis
HAj2: VE(j) is not equal to VEnull}
\item{pval.B1}{p value for testing against the alternative hypothesis
HB1: VE(\code{misscauses}[1]) \eqn{\ge \dots \ge} VE(\code{misscauses}[I]) with strict inequality for some cause in \code{misscauses}}
\item{pval.B2}{p value for testing against the alternative hypothesis
HB2: VE(i) is not equal to VE(j) for at least one pair of i and j in \code{misscauses}}
\item{tgrid}{specifies the time grids for estimating \eqn{\lambda_{0kj}(\cdot)}
, the unspecified cause-specific baseline hazard function
for cause j and the k-th stratum.}
\item{lambda0}{estimates of \eqn{\lambda_{0kj}(\cdot)} using AIPW method}
\item{covariates}{covariates in the cause-specific proportional hazards model.}
\item{trtpos}{the position of the treatment group indicator in the RHS of
the formula object cmprskPHformula.}
\item{VEnull}{the assumed VE value in the null hypothesis.}
\item{diffsigma}{estimates of standard deviation of \eqn{\hat\alpha_j-\hat\alpha_{j-1}}, j=2,...,J.}
\item{cov.alpha}{estimated covariance matrix of \eqn{(\hat\alpha_1,\dots,\hat\alpha_J)}.}
}
\description{
Fit a stratified cause-specific proportional hazards regression model
postulates that the conditional cause-specific hazard function
for cause j for an individual in the k-th stratum with the covariate value
\eqn{z} equals
\deqn{ \lambda_{kj}(t|z)=\lambda_{0kj}(t)\exp(\beta_j^{T}z) }
for j=1,...,J and k=1, ..., K, where \eqn{\lambda_{0kj}(\cdot)}
is an unspecified cause-specific baseline hazard function
for the k-th stratum and K is the number of strata.
}
\details{
Approach 2: Failure endpoint = COVID.
Cause/mark V=0: vaccine-matched genotype AND viral load above minimum threshold;
Cause/mark V=1: vaccine-mismatched genotype AND viral load above minimum threshold;
Cause/mark V=2: viral load below minimum threshold.
In Approach 2, we only care about V=0 and V=1;
V=2 is a competing risk for both V=0 and V=1.

Let R be the indicator of whether all possible data are observed for
a subject; R = 1 if either \eqn{\delta=0} (right-censored) or if
\eqn{\delta=1} (a failure is observed) and the cause is not missing;
and R = 0 otherwise.
Let VL be the viral load and \eqn{h_0} the minimum threshold.
Missing at random (MAR) assumption is valid using Approach 2:
\eqn{P(R=1|\delta=1,T,Z,VL,h_0,V)=P(R=1|\delta=1,T,Z,VL,h_0)}
\eqn{=I(VL<h_0)+P(R=1|\delta=1,T,Z,VL,VL \ge h_0)I(VL\ge h_0)}
}
\references{
(2021+)
}
\author{
Fei Heng
}
