% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{sim2cs}
\alias{sim2cs}
\title{Simulated competing risks data with 2 causes subject to missingness}
\format{
An object of class data.frame
}
\usage{
data(sim2cs)
}
\description{
Simulated data for the package cmprskPH
}
\examples{
data(sim2cs)
}
\references{
Heng et.al. (2021+)
}
\keyword{datasets}
